+++ py/miniconf-mqtt/miniconf/__init__.py
@@ -1,5 +1,4 @@
 """Miniconf Client"""
 
 from .miniconf import Miniconf
-from .version import __version__
 from .discover import discover